Extract out credential_import and helpers
